# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payment API",
    "description": "Payment service for movie tickets",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "payment",
      "description": ""
    }
  ],
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "components": {
    "securitySchemes": {
      "backendToken": {
        "scheme": "bearer",
        "bearerFormat": "token",
        "type": "http",
        "description": "Auth for bearer token",
        "in": "header"
      },
      "redisSessionCookie": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "apiKey",
        "description": "Auth for redis session",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
      "PaymentDTO": {
        "type": "object",
        "properties": {
          "totalPrice": {
            "type": "number",
            "description": "Total Price",
            "example": "20.21"
          },
          "transactionId": {
            "type": "string",
            "description": "Transaction Id",
            "example": "319df989-ec01-48b9-967d-64a68775aaf0"
          }
        },
        "required": [
          "totalPrice",
          "transactionId"
        ]
      }
    }
  },
  "paths": {
    "/booking": {
      "post": {
        "operationId": "payForBooking",
        "summary": "Pay for booking",
        "description": "Paying for booking",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully paid for tickets"
          },
          "400": {
            "description": "Invalid input data"
          },
          "500": {
            "description": "Something went terribly wrong"
          }
        },
        "tags": [
          "payment"
        ],
        "security": [
          {
            "backendToken": []
          }
        ]
      }
    }
  }
}